pyproject.toml
[project]
name = "qrl-trading-bot"
version = "1.0.0"
description = "QRL automated trading bot with risk management"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "ccxt>=4.2.0",
    "pandas>=2.1.0",
    "numpy>=1.26.0",
    "ta>=0.11.0",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
    "SQLAlchemy>=2.0.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "gunicorn>=21.2.0",
    "jinja2>=3.1.2",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "flake8>=6.1.0",
    "black>=23.10.0",
    "mypy>=1.6.0",
]

[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 79
target-version = ['py39', 'py310', 'py311']
include = '\.pyi?$'

[tool.flake8]
max-line-length = 79
extend-ignore = ["E203", "W503"]
exclude = [".git", "__pycache__", "data", "web/templates"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--verbose",
    "--cov=.",
    "--cov-report=term-missing",
]
